preprocess_service/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
google-cloud-storage
numpy
python-multipart
google-cloud-pubsub